+++ Crowdsourcing/src/main/java/tuwien/aic/crowdsourcing/util/MathUtil.java
@@ -6,20 +6,20 @@ import java.util.List;
 
 public abstract class MathUtil {
 
-    public static int[] quartiles(List<Integer> values) {
+    public static float[] quartiles(List<Integer> values) {
         if (values.size() < 3)
             // throw new IllegalArgumentException("This method is not designed to handle lists with fewer than 3 elements.");
-            return new int[0];
+            return new float[0];
 
-        int median = median(values);
+        float median = median(values);
 
         List<Integer> lowerHalf = getValuesLessThan(values, median, true);
         List<Integer> upperHalf = getValuesGreaterThan(values, median, true);
 
-        return new int[] { median(lowerHalf), median, median(upperHalf) };
+        return new float[] { median(lowerHalf), median, median(upperHalf) };
     }
 
-    public static List<Integer> getValuesGreaterThan(List<Integer> values, int limit, boolean orEqualTo) {
+    public static List<Integer> getValuesGreaterThan(List<Integer> values, float limit, boolean orEqualTo) {
         List<Integer> modValues = new ArrayList<Integer>();
 
         for (int value : values)
@@ -29,7 +29,7 @@ public abstract class MathUtil {
         return modValues;
     }
 
-    public static List<Integer> getValuesLessThan(List<Integer> values, int limit, boolean orEqualTo) {
+    public static List<Integer> getValuesLessThan(List<Integer> values, float limit, boolean orEqualTo) {
         List<Integer> modValues = new ArrayList<Integer>();
 
         for (int value : values)
@@ -39,7 +39,7 @@ public abstract class MathUtil {
         return modValues;
     }
 
-    public static int median(List<Integer> values) {
+    public static float median(List<Integer> values) {
         Collections.sort(values);
 
         if (values.size() % 2 == 1)
@@ -48,7 +48,7 @@ public abstract class MathUtil {
             double lower = values.get(values.size() / 2 - 1);
             double upper = values.get(values.size() / 2);
 
-            return (int) ((lower + upper) / 2.0);
+            return (float) ((lower + upper) / 2.0F);
         }
     }
